Removing Framework\Model and Framework\Store
